pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rem-vectordb"
version = "0.2.0"
description = "Official Python SDK for the REM distributed vector database"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "BeClever OÜ", email = "support@getrem.online"}
]
keywords = [
    "vector database",
    "embeddings",
    "semantic search",
    "RAG",
    "AI",
    "machine learning",
    "decentralized",
    "blockchain"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.1.0"]
llamaindex = ["llama-index-core>=0.10.0"]

[project.urls]
Homepage = "https://getrem.online"
Documentation = "https://app.getrem.online/docs"
Repository = "https://github.com/RemNetwork"
"API Platform" = "https://app.getrem.online"
Discord = "https://discord.gg/9ndMQY4PYP"

[tool.setuptools.packages.find]
where = ["."]
include = ["rem*"]
exclude = ["tests*", "examples*"]
